added Effects and synthesize sounds to buttons 1 - 8

+++ Music147_2013/Music147_2013/MUS147Voice_Synth.h
@@ -10,6 +10,6 @@
 
 @interface MUS147Voice_Synth : MUS147Voice
 
--(id)initWithFreq:(UInt32)frequency;
+-(id)initWithFreq:(UInt64)frequency;
 
 @end
